third_party/blink/web_tests/wpt_internal/scheduler/post_task_with_aborted_signal.html

<!doctype html>
<title>Scheduling API: postTask</title>
<link rel="author" title="Scott Haseley" href="mailto:shaseley@chromium.org">
<link rel="help" href="https://wicg.github.io/scheduling-apis/">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<script>
'use strict';

promise_test(t => {
  const controller = new TaskController();
  const signal = controller.signal;
  controller.abort();
  return promise_rejects_dom(t, 'AbortError', scheduler.postTask(() => {}, {signal}));
}, 'Posting a task with an aborted signal rejects with an AbortError');

</script>
